# frontend/openapi/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/v2/applications/{applicationId}/links:
    post:
      operationId: ApplicationLinksController_create
      summary: Create a new link for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkDto"
      responses:
        "201":
          description: Link created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkDto"
      tags: &a17
        - Links
    get:
      operationId: ApplicationLinksController_findAll
      summary: Retrieve links for an application
      description: Get list of links for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
        - name: page
          required: false
          in: query
          description: Numéro de page
          schema:
            minimum: 0
            example: 0
            type: number
        - name: pageSize
          required: false
          in: query
          description: Nombre de résultats par page, 0 pour supprimer la pagination
          schema:
            minimum: 0
            default: 15
            example: 15
            type: number
        - name: sortBy
          required: false
          in: query
          description: Champ utilisé pour le tri
          schema:
            example: label
            type: string
        - name: order
          required: false
          in: query
          description: Ordre de tri
          schema:
            example: asc
            enum:
              - asc
              - desc
            type: string
      responses:
        "200":
          description: List of links for the application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLinkDto"
      tags: *a17
  /api/v2/applications/{applicationId}/links/{id}:
    patch:
      operationId: ApplicationLinksController_update
      summary: Update a link for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
        - name: id
          required: true
          in: path
          description: ID of the link to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkDto"
      responses:
        "200":
          description: Link updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkDto"
      tags: *a17
    delete:
      operationId: ApplicationLinksController_delete
      summary: Delete a link for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
        - name: id
          required: true
          in: path
          description: ID of the link to delete
          schema:
            type: string
      responses:
        "204":
          description: Link deleted successfully
      tags: *a17
  /api/v2/label-sources:
    post:
      operationId: LabelSourceController_create
      summary: Créer une nouvelle source de libellés.
      description: "

        Ce endpoint permet de créer une nouvelle source de libellés alternatifs.


        Information requise :\ 

        - **source** : nom de la source

        \    "
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelSourceDto"
      responses:
        "201":
          description: Source créée avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelSourceDto"
        "403":
          description: Accès refusé - Privilège admin requis
      tags: &a18
        - LabelSources
    get:
      operationId: LabelSourceController_findAll
      summary: Rechercher des sources de libellés alternatifs avec filtre en option.
      parameters:
        - name: page
          required: false
          in: query
          description: Numéro de page
          schema:
            minimum: 0
            example: 0
            type: number
        - name: pageSize
          required: false
          in: query
          description: Nombre de résultats par page, 0 pour supprimer la pagination
          schema:
            minimum: 0
            default: 15
            example: 15
            type: number
        - name: sortBy
          required: false
          in: query
          description: Champ utilisé pour le tri
          schema:
            example: label
            type: string
        - name: order
          required: false
          in: query
          description: Ordre de tri
          schema:
            example: asc
            enum:
              - asc
              - desc
            type: string
        - name: source
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Liste des sources trouvées
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLabelSourceDto"
      tags: *a18
  /api/v2/label-sources/{id}:
    patch:
      operationId: LabelSourceController_update
      summary: Modifier une source de libellés alternatifs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabelSourceDto"
      responses:
        "200":
          description: Source mise à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelSourceDto"
        "403":
          description: Accès refusé - Privilège admin requis
        "404":
          description: Source non trouvée
      tags: *a18
    delete:
      operationId: LabelSourceController_remove
      summary: Supprimer une source de libellés alternatifs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Source supprimée avec succès
        "403":
          description: Accès refusé - Privilège admin requis
        "404":
          description: Source non trouvée
      tags: *a18
  /api/v2/applications/{applicationId}/labels:
    post:
      operationId: LabelsController_create
      summary: Créer un nouveau nom
      description: "

        **Ce endpoint permet de créer un nom complet.**


        Vous devez fournir les informations suivantes :

        - **labelSourceId**: L'id de la source du nom (peut être vide).

        - **value**: Le nom alternatif de l'application.

        \    "
      parameters:
        - name: applicationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelDto"
      responses:
        "201":
          description: Nom alternatif créé avec succès.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelDto"
      tags: &a19
        - Labels
    get:
      operationId: LabelsController_findAllSorted
      summary: Récupérer les noms alternatifs par ID d'application
      description: "

        Ce endpoint permet de récupérer la liste de tous les noms alternatifs
        d'une application en fonction de son identifiant unique.


        Le paramètre **applicationId** doit être fourni dans l'URL.

        \    "
      parameters:
        - name: applicationId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Liste des noms alternatifs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabelDto"
      tags: *a19
  /api/v2/applications/{applicationId}/labels/{id}:
    patch:
      operationId: LabelsController_update
      summary: Mettre à jour un nom existant
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
        - name: id
          required: true
          in: path
          description: ID du nom
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelDto"
      responses:
        "200":
          description: Nom alternatif mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabelDto"
      tags: *a19
    delete:
      operationId: LabelsController_delete
      summary: Supprimer un nom alternatif
      description: " Ce endpoint permet de supprimer un nom alternatif existant.\ 

        \    Vous devez fournir l'identifiant du nom alternatif dans l'URL.

        \    "
      parameters:
        - name: applicationId
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Nom alternatif supprimé avec succès
      tags: *a19
...
info:
  title: API Référentiel Applications
  description: API pour la gestion des applications
  version: "2.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/auth
          tokenUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          refreshUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          scopes:
            openid: OpenID scope
            profile: Profile scope
      description: OAuth2 authentication using OIDC
    api_key:
      type: apiKey
      in: header
      name: x-refapp-token
      description: Token authentication
  schemas:
...
    ExternalRessourceType:
      type: string
      description: Type of the external resource
      enum:
        - documentation
        - supervision
        - service
        - audience_metrics
        - main_service
    CreateLinkDto:
      type: object
      properties:
        link:
          type: string
          description: The URL of the link
        type:
          $ref: "#/components/schemas/ExternalRessourceType"
        description:
          type: string
          description: Optional description of the link
      required:
        - link
        - type
    LinkDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the link
        link:
          type: string
          description: The URL of the link
        type:
          type: string
          enum:
            - documentation
            - supervision
            - service
            - audience_metrics
            - main_service
          description: Type of the external resource
        description:
          type: string
          description: Description of the link
      required:
        - id
        - link
        - type
    PaginatedLinkDto:
      type: object
      properties:
        results:
          description: Array of results
          type: array
          items:
            $ref: "#/components/schemas/LinkDto"
        total:
          type: number
          description: Total number of items
          example: 100
      required:
        - results
        - total
    UpdateLinkDto:
      type: object
      properties:
        link:
          type: string
          description: The URL of the link
        type:
          $ref: "#/components/schemas/ExternalRessourceType"
        description:
          type: string
          description: Optional description of the link
    CreateLabelSourceDto:
      type: object
      properties:
        source:
          type: string
          description: Source name
          example: CODE_PAI
      required:
        - source
    LabelSourceDto:
      type: object
      properties:
        source:
          type: string
          description: Source name
          example: CODE_PAI
        id:
          type: string
          description: Unique identifier of the label source
          example: 5708d232-8338-4abf-8f38-8370acc89497
      required:
        - source
        - id
    PaginatedLabelSourceDto:
      type: object
      properties:
        results:
          description: Array of results
          type: array
          items:
            $ref: "#/components/schemas/LabelSourceDto"
        total:
          type: number
          description: Total number of items
          example: 100
      required:
        - results
        - total
    UpdateLabelSourceDto:
      type: object
      properties:
        source:
          type: string
          description: Source name
          example: CODE_PAI
      required:
        - source
    CreateLabelDto:
      type: object
      properties:
        labelSourceId:
          type: string
          example: 5708d232-8338-4abf-8f38-8370acc89497
          description: ID de la source du label
        value:
          type: string
          example: short-app-name
          description: Value of the label
      required:
        - value
    LabelDto:
      type: object
      properties:
        labelSourceId:
          type: string
          example: 5708d232-8338-4abf-8f38-8370acc89497
          description: ID de la source du label
        value:
          type: string
          example: short-app-name
          description: Value of the label
        id:
          type: string
          example: 5708d232-8338-4abf-8f38-8370acc89497
          description: ID of the label
        applicationId:
          type: string
          example: 5708d232-8338-4abf-8f38-8370acc89497
          description: ID of the application to which the label belongs
      required:
        - value
        - id
        - applicationId
...
security:
  - oauth2: []
  - api_key: []